Use int64 because Firestore doesn't support int32

+++ backend/dto.go
@@ -12,10 +12,10 @@ type dto struct {
 	Name                        string    `firestore:"name"`
 	URL                         string    `firestore:"url"`
 	Owners                      []string  `firestore:"owners"`
-	RedirectCount28Days         int32     `firestore:"redirect_count_28days"`
-	RedirectCount7Days          int32     `firestore:"redirect_count_7days"`
+	RedirectCount28Days         int64     `firestore:"redirect_count_28days"`
+	RedirectCount7Days          int64     `firestore:"redirect_count_7days"`
 	RedirectCountCalculatedDate time.Time `firestore:"redirect_count_calculated_date"`
-	DailyRedirectCounts         [28]int32 `firestore:"daily_redirect_counts"`
+	DailyRedirectCounts         [28]int64 `firestore:"daily_redirect_counts"`
 	CreatedAt                   time.Time `firestore:"created_at"`
 	UpdatedAt                   time.Time `firestore:"updated_at"`
 }
